Use valid directory for Sphinx’s PID file
Sphinx refuses to start if it can’t create its PID file.
So we need to make sure the sphinxsearch user always has
write permissions in the directory holding the PID file.

On Stretch, /var/run is a symbolic link to /run, which is
mounted on tmpfs. Therefore, we need to setup the
directory /run/sphinxsearch/ using a tmpfiles.d definition.