Add one more bit to the private network subnet
It looks like Windows refuses to configure a network interface
with an IP that corresponds to the first address (all zeros)
or the last address (all ones) available in the subnet.

A netmask of 255.255.255.252 allows two bits of address space,
so four addresses in total, of which two cannot be used by
Windows, which leaves one for the host and one for the guest.