Prevent sql_mode from changing
Upgrading to Debian Buster will update MariaDB from 10.1.38
to 10.3.22. However, MariaDB changed the default value of
sql_mode starting from 10.2.4 [1]. This new value breaks
a lot of things in Tatoeba because it turns many MySQL
warnings into fatal errors. So we need to use the old
default value for now.

[1] https://mariadb.com/kb/en/sql-mode/#defaults